+++ guillotina_amqp/utils.py
@@ -202,14 +202,13 @@ class TimeoutLock(object):
         return await self.acquire(ttl)
 
 
-def measure(metric, value, **labels):
-    try:
-        import prometheus_client
-    except ImportError:
+def metric_measure(metric, value, labels=None):
+    if not metric:
         # Don't measure if prometheus client is not installed
         return
 
     try:
+        labels = labels or {}
         metric.labels(**labels).observe(value)
     except Exception:
         logger.error(f'Failed to measure metric', exc_info=True)